DEV: Do not include www in github url (#17)
Why did I notice? Pfaffmanager uses `/admin/plugins.json` to see what plugins are installed and uses that URL as part of a regex to remove the plugin from `app.yml`. It took me a while to figure out why it wouldn't delete this one! I might have tweaked my code, but this seems to be the only plugin that I've noticed this in. And github doesn't use www.